+++ build_modules/lib/build_modules.dart
@@ -3,8 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 export 'src/errors.dart' show MissingModulesException;
-export 'src/kernel_builder.dart'
-    show KernelSummaryBuilder, kernelSummaryExtension, multiRootScheme;
+export 'src/kernel_builder.dart' show KernelBuilder, multiRootScheme;
 export 'src/meta_module_builder.dart'
     show MetaModuleBuilder, metaModuleExtension;
 export 'src/meta_module_clean_builder.dart'